backend/requirements.txt
fastapi
uvicorn
pydantic
boto3
python-docx
pdfplumber
sentence-transformers
qdrant-client
httpx
python-dotenv
latex